feat: Implement procedural side-scrolling shooter game
This commit introduces a complete, playable side-scrolling shooter game inspired by "Gradius," built with Pygame.

Key features implemented:
- Player spaceship with movement (arrow keys) and shooting (spacebar).
- Enemies that spawn from the right and move left.
- Continuous side-scrolling environment.
- Collision detection:
    - Player-Enemy: Game Over.
    - Player-Obstacle: Game Over.
    - Bullet-Enemy: Enemy destroyed, score increases.
- Procedural Stage Generation:
    - Each stage has a randomly determined duration (45-90 seconds).
    - Enemy types (speed) and spawn frequency are randomized and scale with stage progression.
    - Obstacles (static grey rectangles) are randomly placed, with their count increasing in later stages.
- Stage Goal: Survive for the duration of the stage.
- Progression: Upon survival, a new, more difficult stage is generated.
- HUD: Displays Score, Remaining Stage Time, and Current Stage Number.
- Technical Specifications:
    - Uses Pygame.
    - All game elements (player, enemies, bullets, obstacles) are represented by simple geometric shapes (pygame.Rect) with specified colors. No external assets are used.
    - Code is structured with classes for Player, Enemy, Bullet, Obstacle, and a main Game class.
- Comments: Added throughout the code to explain logic, especially procedural generation and difficulty scaling.
- User Feedback: Incorporated a fix for potential float issue in random.randint for obstacle height calculation (`int(SCREEN_HEIGHT * 0.6)`).

The game is contained in a single Python file (`gradius_game.py`).
A planned refinement for adding a bullet cooldown was not implemented due to turn limitations.

+++ gradius_game.py
@@ -215,7 +215,7 @@ class Game:
         for i in range(num_obstacles):
             # Obstacle dimensions are randomized.
             obs_width = random.randint(30, 120)
-            obs_height = random.randint(30, SCREEN_HEIGHT * 0.6) # Height up to 60% of screen
+            obs_height = random.randint(30, int(SCREEN_HEIGHT * 0.6)) # Height up to 60% of screen
             # Obstacle X position is staggered off-screen to the right.
             obs_x = SCREEN_WIDTH + i * (SCREEN_WIDTH / num_obstacles) + random.randint(50, 300)
             max_y = SCREEN_HEIGHT - obs_height # Ensure obstacle is fully within screen vertically
